Move wait_until() into supervisord.py
Hideous fix for a packaging problem: we don't want the API to depend on
ctxrunner dependencies but right now it does because:
- The API imports the DB etc from `damnit.backend`
- `damnit.backend.__init__` imports stuff from `supervisord.py`
- `supervisord.py` imports `wait_until()` from `utils.py`
- `utils.py` imports stuff from `ctxrunner`
- And thus the API depends on `ctxrunner` and all its dependencies

Hence by moving `wait_until()` directly into `supervisord.py` we avoid the
dependency on `ctxrunner`.
